Add interface for retrieving alerting config from charmed AlertManager (#14)
This commit allows the charm to retrieve a configuration string from the
alertmanager charm. The configuration string is a JSON string that this
charm converts to a dict that is then inserted to another dict that
ultimately gets serialized into a YAML-formatted prometheus config file.

While this charm is agnostic to the contents of the JSON string, in
AlertManager's current iteration, it uses the `kube_sd_configs` method
so that Prometheus can quickly detect changes in the number of
AlertManager pods without waiting for any Juju events.